Move page-footer to be inside page-content
Because of the weird multiple levels of scrollbars, moving the
page-footer into page-content fixes a problem where the footer would
sorta hang in this large blank space at the end of the page and cause a
weird break in scrollable area. When inside page-content the footer
is part of the scrollable area.