Introduce Project.IsFaved
Move away from a simple bool to represent the favorite status of a
project, and instead use a sum type to better represent UI states like
"just faved", which is useful for animating when clicking to favorite.

Also add the Unknown variant to handle favorite status for logged out
users.