Syntax: Prefer HQ.NameOnly when we have an FQN
`HashQualified.NameOnly` allows us to constructor urls based on the name.
References in Syntax, like `TermReference` now include a full name (the
`fqn` field), so when that is present, convert it to a `NameOnly`.

It is true that we loose data by not using `HashQualified`, but
currently that variant can't be correctly encoded to the URL, so prefer
the `NameOnly` variant.

Could we have done this in the consuming app (like share-ui)? Yes, but
not in a backwards compatible way. The `HashQualified` was previously
used for the display name (aka not a full name), so if this is to
successfully be deployed before the backend is ready for ucm and share
we would be linking to referenced based on too short of a name.